# Improving performance of AMR with p4est