--- a/src/mesh/mesh_io.jl
+++ b/src/mesh/mesh_io.jl
@@ -162,7 +162,9 @@ function save_mesh_file(mesh::P4estMesh, output_directory, timestep=0)
     attributes(file)["p4est_file"] = p4est_filename
 
     file["tree_node_coordinates"] = mesh.tree_node_coordinates
-    file["nodes"] = mesh.nodes
+    file["nodes"] = Vector(mesh.nodes) # the mesh might use custom array types
+                                       # to increase the runtime performance
+                                       # but HDF5 can only handle plain arrays
     file["boundary_names"] = mesh.boundary_names .|> String
   end