--- a/src/solvers/dg_p4est/containers.jl
+++ b/src/solvers/dg_p4est/containers.jl
@@ -418,10 +418,15 @@ end
 
 # Let p4est iterate over all interfaces and execute the C function iter_face_c
 function iterate_faces(mesh::P4estMesh, iter_face_c, user_data)
+  if user_data isa AbstractArray
+    user_data_ptr = pointer(user_data)
+  else
+    user_data_ptr = pointer_from_objref(user_data)
+  end
   GC.@preserve user_data begin
     p4est_iterate(mesh.p4est,
                   C_NULL, # ghost layer
-                  pointer(user_data),
+                  user_data_ptr,
                   C_NULL, # iter_volume
                   iter_face_c, # iter_face
                   C_NULL) # iter_corner