@efaulhaber 2021-06-10
Thank you very much! And I appreciate the thorough guide on how to approach performance optimizations like that.

~~Quick~~ Longer than expected question before I look at the code: Have you tried the native `@profile` output of the newer `julia-vscode` versions?
Running your benchmark code from step 1 in vscode without `using ProfileView` first gives me this output:

![grafik](https://user-images.githubusercontent.com/44124897/121605077-0a0ecf00-ca4c-11eb-9236-0c96e541e75e.png)

There's also an option to show a flame graph:

![grafik](https://user-images.githubusercontent.com/44124897/121605122-21e65300-ca4c-11eb-9934-e78b733c56c2.png)

I can see that a lot of time is spent in `iterate_faces` and the `mul!` lines of `calc_jacobian_matrix!`, but I can't see anything in the flame graph.
These two seem to be reversed (compared to `ProfileView`) in that they collect all calls to the low level functions (like `LinearAlgebra._generic_matmatmul!`), which are shown by `ProfileView` in the top rows, and display them in the first level. Then, I can find out which of my functions called these by expanding the entries.
I find `ProfileView`'s output much more useful. Here, it immediately shows that `iterate_faces` is responsible for less than half of the time, while `calc_jacobian_matrix!` and `calc_node_coordinates` are responsible for more than half. This can't be seen in vscode because the latter calls are scattered across different low-level function calls.

I'd like to have the first picture in reverse, like a quantitative version of the flame graph by `ProfileView`. I like these quantitative representations a lot better than the flame graph that only shows me useful information when I hover over it. Do you happen to know if there's any way to show such a quantitative representation of `ProfileView`'s flame graph?