@ranocha 2021-06-11
> Have you tried the native `@profile` output of the newer `julia-vscode` versions?

No, I have never tried that.

> Do you happen to know if there's any way to show such a quantitative representation of `ProfileView`'s flame graph?

What exactly do you mean by that?