@sloede 2021-06-11
Thanks for the timings! The p4est AMR now looks very competitive in comparison to the the TreeMesh. If we exclude the solver-specific parts from the `AMR` timer, and cut the time for `rebalance` in half, we have 7.5ms for TreeMesh and 15.9ms for P4estMesh, just for the mesh part. Given that p4est is much more capable in terms of parallelization, I think this is a very good result already!